fix payout input body type , and custom result for 500

+++ backend/Booking/src/Modules/Booking.Modules.Mentorships/Features/Payout/User/Request/Payout.cs
@@ -10,7 +10,7 @@ namespace Booking.Modules.Mentorships.Features.Payout.User.Request;
 
 public class Payout : IEndpoint
 {
-    public record Request(int Amount);
+    public record Request(decimal Amount);
 
     public void MapEndpoint(IEndpointRouteBuilder app)
     {
