+++ backend/Booking/src/Modules/Booking.Modules.Mentorships/Features/Payout/User/Request/PayoutCommand.cs
@@ -2,4 +2,4 @@ using Booking.Common.Messaging;
 
 namespace Booking.Modules.Mentorships.Features.Payout.User.Request;
 
-public record PayoutCommand(int UserId, int Amount) : ICommand;
+public record PayoutCommand(int UserId, decimal Amount) : ICommand;
